packages/infra/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "infra"
version = "0.1.0"
description = "AWS CDK infrastructure for OT Threat Research Assistant"
requires-python = ">=3.12,<3.13"
dependencies = [
    "aws-cdk-lib>=2.130",
    "constructs>=10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/infra"]
